Move search movie cmps to its own dir
