+++ pages/movies-search/[searchterm].tsx
@@ -2,8 +2,8 @@ import withApollo from '../../lib/apollo'
 import { useRouter } from 'next/router'
 import { useQuery } from '@apollo/react-hooks'
 import { NoResults } from '../../components/NoResults'
-import { Pagination } from '../../components/Pagination'
-import { MovieList } from '../../components/MovieList'
+import { Pagination } from '../../components/movies-search/Pagination'
+import { MovieList } from '../../components/movies-search/MovieList'
 import { TopBar } from "../../components/TopBar";
 import { CONST } from '../../constants'
 import { GET_MOVIES } from '../../schemas'
